requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
python-telegram-bot==20.7
python-dotenv==1.0.0
Flask-WTF==1.2.1
WTForms==3.1.1
requests==2.31.0

